move src package to ergo
