+++ ergo/ergo_cli.py
@@ -5,13 +5,13 @@ import os
 import yaml
 from colors import color
 
-from src.amqp_invoker import AmqpInvoker
-from src.config import Config
-from src.flask_http_invoker import FlaskHttpInvoker
-from src.function_invocable import FunctionInvocable
-from src.http_invoker import HttpInvoker
-from src.schematic import graph as ergograph
-from src.version import get_version
+from ergo.amqp_invoker import AmqpInvoker
+from ergo.config import Config
+from ergo.flask_http_invoker import FlaskHttpInvoker
+from ergo.function_invocable import FunctionInvocable
+from ergo.http_invoker import HttpInvoker
+from ergo.schematic import graph as ergograph
+from ergo.version import get_version
 
 
 def format_date(sec: float) -> str:
