+++ ergo/flask_http_invoker.py
@@ -5,8 +5,8 @@ from typing import List
 
 from flask import Flask, request  # , abort
 
-from src.http_invoker import HttpInvoker
-from src.types import TYPE_PAYLOAD
+from ergo.http_invoker import HttpInvoker
+from ergo.types import TYPE_PAYLOAD
 
 
 class FlaskHttpInvoker(HttpInvoker):
